refactor(file-structure): move project_types to root dir
- Move get_tutorial_task_property to project_types/store
