+++ apps/tutorial/serializers.py
@@ -7,29 +7,11 @@ from rest_framework import serializers
 
 from apps.common.serializers import DrfContextType, UserResourceSerializer
 from apps.project.models import ProjectTypeEnum
-from apps.tutorial.project_types.tile_map_service.compare import tutorial as compare_tutorial
-from apps.tutorial.project_types.tile_map_service.completeness import tutorial as completeness_tutorial
-from apps.tutorial.project_types.tile_map_service.find import tutorial as find_tutorial
-from apps.tutorial.project_types.validate import tutorial as validate_tutorial
+from project_types.store import get_tutorial_task_property
 from utils.common import clean_up_none_keys
 from utils.graphql.drf import handle_pydantic_validation_error
 
 from .models import Tutorial, TutorialInformationPage, TutorialInformationPageBlock, TutorialScenarioPage, TutorialTask
-
-
-# FIXME(tnagorra): Move this to store
-def get_tutorial_task_property(project_type: ProjectTypeEnum | None):
-    if project_type is None:
-        return None
-    if project_type == ProjectTypeEnum.COMPARE:
-        return ("compare", compare_tutorial.CompareTutorialTaskProperty)
-    if project_type == ProjectTypeEnum.FIND:
-        return ("find", find_tutorial.FindTutorialTaskProperty)
-    if project_type == ProjectTypeEnum.VALIDATE:
-        return ("validate", validate_tutorial.ValidateTutorialTaskProperty)
-    if project_type == ProjectTypeEnum.COMPLETENESS:
-        return ("completeness", completeness_tutorial.CompletenessTutorialTaskProperty)
-    typing.assert_never(project_type)
 
 
 class TutorialTaskSerializerContextType(DrfContextType):
