+++ project_types/tile_map_service/compare/project.py
@@ -1,8 +1,8 @@
 import typing
 
 from apps.project.models import Project, ProjectTask, ProjectTaskGroup, ProjectTypeEnum
-from apps.project.project_types.tile_map_service.base import project as base_project
 from main.bulk_managers import BulkCreateManager
+from project_types.tile_map_service.base import project as base_project
 from utils.geo import tile_functions, tile_grouping
 from utils.geo.tile_server.models import TileServerConfig
 from utils.geo.tile_server.tile_server import AvailableTileServerTypeAlias, get_tile_server
