+++ project_types/tile_map_service/compare/tutorial.py
@@ -1,4 +1,4 @@
-from apps.tutorial.project_types.tile_map_service.base import tutorial as tile_map_service_tutorial
+from project_types.tile_map_service.base import tutorial as tile_map_service_tutorial
 
 
 class CompareTutorialTaskProperty(tile_map_service_tutorial.TileMapServiceTutorialTaskProperty): ...
